Simulate on the entire mesh specified by the user.
Context : the user gives a mesh as input of OpenTURNSFMUPointToFieldFunction. The FMU output is interpolated on this mesh.

Issue : when the mesh last value is larger than the FMU default simulation time, OpenTURNSFMUPointToFunction output is filled with the FMU last value. The output is thus evolving from time=0 to time=FMU_default_simulation_time; then it is constant.

Improvement : the mesh last value now specifies the FMU simulation
time.